refactor: rename Recent Changes to Changelog, move API under /api/pages/
Rename the /recent-changes route to /changes, display "Changelog" in the
nav and page title, and move the backend API from /api/recent-changes/ to
/api/pages/changes/ per the page-oriented API convention. Extract shared
schemas into provenance/schemas.py to avoid circular imports between
api.py and the new page_endpoints.py.

UX improvements: replace UserBadge with inline "By user" byline,
remove "Include data imports" toggle, drop time from dates older than
7 days, use full weekday names in SmartDate.
